arista-mcp-server/pyproject.toml
[project]
name = "arista-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi-mcp>=0.3.4",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
]
